backend/requirements.txt
certifi==2020.4.5.1
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
SQLAlchemy==1.3.16
Werkzeug==1.0.1
wincertstore==0.2
python-dotenv
psycopg2-binary
googlesearch-python
numpy
pandas
python-dateutil
requests
beautifulsoup4
python-whois
flask-cors
Flask-Mail
xgboost
joblib
scikit-learn